refactor(wordpress): centralize shared profiler defaults
